+++ src/aideas/app/run_context.py
@@ -5,8 +5,7 @@ from .action.action_result import ActionResult
 from .action.variable_parser import replace_all_variables
 from .config import AppConfig, RunConfig
 from .env import Env
-from .result.result_set import AgentResultSet, ElementResultSet, StageResultSet
-
+from .result.result_set import AgentResultSet, StageResultSet, ElementResultSet
 
 CURRENT_URL = 'current_url'
 CURRENT_ELEMENT = 'current_element'
@@ -49,59 +48,21 @@ class RunContext:
         return config
 
     def add_action_result(self, result: ActionResult) -> 'RunContext':
-        agent_name = result.get_action().get_agent_name()
-        stage_id = result.get_action().get_stage_id()
-        new_stage = False
-        new_stage_element = False
-        stage_result_set: StageResultSet = self.get_stage_results(agent_name, None)
-        if stage_result_set is None:
-            new_stage = True
-            stage_result_set = StageResultSet()
-        element_result_set: ElementResultSet = stage_result_set.get(stage_id, None)
-        if element_result_set is None:
-            new_stage_element = True
-            element_result_set = ElementResultSet()
-        element_result_set.add(result)
-        if new_stage_element:
-            stage_result_set.set(stage_id, element_result_set)
-        if new_stage:
-            self.__result_set.set(agent_name, stage_result_set)
+        self.__result_set.add_action_result(result)
         return self
-
-    def get_action_results(self,
-                           agent_name: str,
-                           stage_id: str,
-                           element_name: str) -> list[ActionResult]:
-        element_result_set: ElementResultSet = self.get_element_results(agent_name, stage_id)
-        return [] if element_result_set is None else element_result_set.get(element_name, [])
-
-    def get_action_result(self,
-                          agent_name: str,
-                          stage_id: str,
-                          result_if_none: Union[ActionResult, None] = ActionResult.none()) \
-            -> ActionResult:
-        result_set: ElementResultSet = self.get_element_results(agent_name, stage_id)
-        if result_set.is_empty():
-            return result_if_none
-        key: str = next(iter(result_set.keys()))
-        action_result_list: list[ActionResult] = result_set.get(key, [])
-        action_result = None if len(action_result_list) == 0 else action_result_list[0]
-        return action_result if action_result is not None else result_if_none
 
     def get_element_results(self,
                             agent_name: str,
                             stage_id: str,
                             result_if_no: Union[ElementResultSet, None] = ElementResultSet.none()) \
             -> ElementResultSet:
-        stage_result_set: StageResultSet = self.get_stage_results(agent_name)
-        return result_if_no if stage_result_set is None else (
-            stage_result_set.get(stage_id, result_if_no))
+        return self.__result_set.get_element_results(agent_name, stage_id, result_if_no)
 
     def get_stage_results(self,
                           agent_name: str,
                           result_if_none: Union[StageResultSet, None] = StageResultSet.none()) \
             -> StageResultSet:
-        return self.__result_set.get(agent_name, result_if_none)
+        return self.__result_set.get_stage_results(agent_name, result_if_none)
 
     def get_agent_names(self) -> list[str]:
         return [e for e in self.__agent_names]